refactor(shared): split db/schema.ts by bounded context, restore 800 cap
Split the 911-line db/schema.ts god-file into eight per-context sibling
modules (auth, location, fleet, pricing, booking, payment, messaging,
notification), re-exported from db/schema.ts so every
@kuruma/shared/db/schema import and drizzle-kit's snapshot discovery stay
stable. Mirrors the existing db/ fragment convention (add-on, regions,
provider-access, renter-documents, booking-types).

Each module colocates its tables, pgEnums, and the derived enum
string-union types. The FK graph is a clean DAG (no import cycles);
composite-FK targets are imported from upstream context modules.

Dropped the lint-file-size special-case (cap=1000, soft-warn=800) for
schema.ts so the standard 800 hard cap now applies to the barrel; updated
its test accordingly.

Proven behavior-preserving: db:generate is a no-op (drizzle snapshot
byte-identical), db:verify 5/5, db:migrate clean on a fresh DB, typecheck
shared/api/web, shared suite 542/542, lint:fk-indexes 49/49.

Closes #725
